+++ src/common/types/network.ts
@@ -2,11 +2,10 @@ import { ChainID } from '@stacks/transactions';
 
 export enum NetworkModes {
   Testnet = 'testnet',
-  Devnet = 'devnet',
   Mainnet = 'mainnet',
 }
 
-export type NetworkMode = NetworkModes.Mainnet | NetworkModes.Testnet | NetworkModes.Devnet;
+export type NetworkMode = NetworkModes.Mainnet | NetworkModes.Testnet;
 
 export interface Network {
   label: string;